made sigle story view work with sockets so user comments appear on appropriate stories in real time
