+++ client/components/singleStoryView.js
@@ -8,7 +8,7 @@ import { fetchStoryThunk, } from '../store';
 const SingleComment = (props) => {
 return (
     <div>
-        <h4>{props.commentObj.user.displayName}</h4>
+       {/* <h4>{props.commentObj.user.displayName}</h4>*/}
         <p>{props.commentObj.content}</p>
     </div>
 )
@@ -20,23 +20,24 @@ class SingleStory extends React.Component{
         this.state = {}
     }
     componentDidMount(){
-        this.props.fetchStory(this.props.match.params.id)
+        this.props.fetchStory(this.props.id)
     }
     render(){
     return (
         <div>
-        <p>by {this.props.story.user.displayName}</p>
-            <MediaPlayer />
+        <p>by {Object.keys(this.props.story).length ? this.props.story.user.displayName : 'Loading'}</p>
+           <MediaPlayer />
             <h2>Comments:</h2>
             {Object.keys(this.props.currentUser).length ? <CommentForm /> : <p>Sign in to leave comments</p>}
-            {this.props.allComments.map(commentObj => <SingleComment key={commentObj.id} commentObj={commentObj} />)}
+            {Object.keys(this.props.story).length ? this.props.allComments.map(commentObj => <SingleComment key={commentObj.id} commentObj={commentObj} />) : 'Loading Comments'}
         </div>
     )}
 }
-const mapState = (state) => ({
+const mapState = (state, ownProps) => ({
     currentUser: state.user,
     allComments: state.story.comments,
     story: state.story,
+    id: ownProps.match.params.id,
 });
 const mapDispatch = (dispatch) => ({
     fetchStory: function (id){
