+++ client/index.js
@@ -8,7 +8,7 @@ import App from './app'
 
 // establishes socket connection
 import './socket'
-import './style.css'
+import '../public/style.css'
 ReactDOM.render(
   <Provider store={store}>
     <Router history={history}>
