+++ client/store/comments.js
@@ -1,12 +1,8 @@
-import Axios from 'axios';
-import socket from '../socket';
 
 //action type
 
 const NEW_COMMENT_CONTENT = 'NEW_COMMENT_CONTENT';
 const CLEAR_COMMENT_FORM = 'CLEAR_COMMENT_FORM';
-const GET_COMMENTS = 'GET_COMMENTS';
-const GET_NEW_COMMENT = 'GET_NEW_COMMENT';
 
 //action creators
 
@@ -19,28 +15,9 @@ export const clearCommentForm = () => ({
     type: CLEAR_COMMENT_FORM,
 })
 
-export const getComments = (comments) => ({
-    type: GET_COMMENTS,
-    comments,
-})
-
-export const getNewComment = (comment) => ({
-    type: GET_NEW_COMMENT,
-    comment,
-})
-//thunk
-export const getNewCommentThunk = (comment) => (dispatch) => {
-    Axios.post(`/story/${comment.storyId}`, comment)
-    .then(newComment => {
-        dispatch(getNewComment(newComment))
-        socket.emit('new-comment', newComment)
-    })
-}
-
 //initial state
 let initialState = {
     content: '',
-    comments: [],
 }
 
 //reducer
@@ -50,11 +27,7 @@ export default function (state = initialState, action){
             return {...state, content: action.content, }
         case CLEAR_COMMENT_FORM:
             return {...state, content: '', }
-        case GET_COMMENTS:
-            return {...state, comments: [...action.comments, ], }
-        case GET_NEW_COMMENT:
-            return {...state, comments: [...state.comments, action.comment, ], }
-        default:
+       default:
             return state
     }
 }
