+++ client/store/story.js
@@ -1,8 +1,10 @@
 import Axios from 'axios';
+import socket from '../socket';
 
 //action type
 
 const FETCH_STORY = 'FETCH_STORY';
+const GET_NEW_COMMENT = 'GET_NEW_COMMENT';
 
 
 //action creators
@@ -12,15 +14,30 @@ export const fetchStory = story => ({
     story,
 })
 
+export const getNewComment = (comment) => ({
+    type: GET_NEW_COMMENT,
+    comment,
+})
 
 //initial state
 let initialState = {}
 
 //thunk creator
 export const fetchStoryThunk = (storyId) => (dispatch) => {
-    Axios.get(`/story/${storyId}`)
+    console.log(storyId)
+    Axios.get(`/api/story/${storyId}`)
+    .then(res => res.data)
     .then(storyObj => {
-        dispatch(fetchStory(storyObj))
+        dispatch(fetchStory(storyObj[0]))
+    })
+}
+
+export const getNewCommentThunk = (comment) => (dispatch) => {
+    Axios.post(`/api/story/${comment.storyId}`, comment)
+    .then(res => res.data)
+    .then(newComment => {
+        dispatch(getNewComment(newComment))
+        socket.emit('new-comment', newComment)
     })
 }
 
@@ -28,7 +45,14 @@ export const fetchStoryThunk = (storyId) => (dispatch) => {
 export default function (state = initialState, action){
     switch (action.type){
         case FETCH_STORY:
-            return {...state, story: action.story, }
+            return {...action.story, }
+        case GET_NEW_COMMENT:
+            if (action.comment.storyId === state.id){
+            return {...state, comments: [...state.comments, action.comment, ], }
+        }
+        else {
+            return {...state, }
+        }
         default:
             return state
     }
